+++ src/Model/Model-firestore.js
@@ -1,0 +1,8 @@
+const createData = (post, correo) => firebase.firestore().collection('Posts').add({
+  text: post,
+  email: correo,
+});
+
+export {
+  createData,
+};
